backend/pyproject.toml
[project]
name = "cloumask-backend"
version = "0.1.0"
description = "Python sidecar for Cloumask - CV processing and LangGraph agent"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Cloumask Team" }
]
keywords = ["computer-vision", "llm", "agent", "fastapi", "tauri"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]
agent = [
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
]
cv = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "opencv-python>=4.10.0",
    "numpy>=1.26.0",
    "pillow>=10.4.0",
]
all = [
    "cloumask-backend[dev,agent,cv]",
]

[project.scripts]
cloumask-backend = "backend.api.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PTH"]
ignore = ["E501", "B008"]  # B008: FastAPI Query/Path/Body in defaults is acceptable

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
